pyproject.toml
[tool.poetry]
name = "async-django-demo"
version = "0.1.0"
description = ""
authors = ["Yuri Masalov <masalov.09@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
django = "^5.1.7"
daphne = "^4.1.2"
adrf = "^0.1.9"
djangorestframework = "^3.15.2"
django-compressor = "^4.5.1"
whitenoise = "^6.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
